Updated this to the best of my knowledge
I wasnt quite sure how to translate series as I'm not sure to what that refers to. If u have any questions you can try to contact me on discord  (punki#8284)and we can see if we can work things out. I'm also on the peomap server in any case, so i will probably any new anouncement. try not to contact on github as I will likely will not be only here very often.

Best Regards